Post-review updates - used named placeholders for string interpolation - removed re-export of wagtail modeladmin - added more comments in modeladmin views - refactored modeladmin as an optional separate sub-app - moved modeladmin submit view to modeladmin.views - moved tests to modeladmin/tests.py
