+++ wagtail_localize/modeladmin/__init__.py
@@ -1,1 +1,10 @@
-from .options import *  # noqa
+from django import VERSION as DJANGO_VERSION
+
+
+if DJANGO_VERSION >= (3, 2):
+    # The declaration is only needed for older Django versions
+    pass
+else:
+    default_app_config = (
+        "wagtail_localize.modeladmin.apps.WagtailLocalizeModelAdminAppConfig"
+    )
