+++ wagtail_localize/modeladmin/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailLocalizeModelAdminAppConfig(AppConfig):
+    name = "wagtail_localize.modeladmin"
+    label = "wagtail_localize_modeladmin"
+    verbose_name = "Wagtail localize modeladmin"
